+++ composeApp/src/commonMain/kotlin/scenes/mainscreen/MainScreen.kt
@@ -1,6 +1,6 @@
 package scenes.mainscreen
 
-import AppColors
+import sharedData.AppColors
 import androidx.compose.animation.core.animateDpAsState
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
