+++ composeApp/src/commonMain/kotlin/scenes/onboarding/OnboardingScreen.kt
@@ -1,6 +1,6 @@
 package scenes.onboarding
 
-import AppColors
+import sharedData.AppColors
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
