+++ composeApp/src/commonMain/kotlin/scenes/profilescreen/ProfileScreen.kt
@@ -1,6 +1,6 @@
 package scenes.profilescreen
 
-import AppColors
+import sharedData.AppColors
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
