+++ composeApp/src/commonMain/kotlin/ui/DevicesBubbleContent.kt
@@ -1,6 +1,6 @@
-package sharedData
+package ui
 
-import AppColors
+import sharedData.AppColors
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
